Add MapViewModel
