+++ UsefulMap/UsefulMap/LocationRequest/Model/LocationManager.swift
